+++ web/src/nav/NavLinks.tsx
@@ -96,6 +96,7 @@ export class NavLinks extends React.PureComponent<Props> {
                     extensionsController={this.props.extensionsController}
                     platformContext={this.props.platformContext}
                     toggleVisibilityKeybinding={this.props.keybindings.commandPalette}
+                    autoFocus={true}
                     location={this.props.location}
                 />
                 {this.props.authenticatedUser && (